A big commit introducing `nvim` setup and migration
This is the initial attempt at migrating my current vim setup to the
nvim ecosystem, including:

- All current key mappings
- `vim.set.*` options
- Plugin manager and requisite plugins

I'm likely missing a few fringe settings, but overall, I'm happy with
the current state my setup is in after taking them for a few short test
drives. It's not the most complex setup in the world, but I do
anticipate modifications overtime as I learn more about the power of Lua
at my disposal.